# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MediCare HMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /lab/orders:
    get:
      operationId: listLabOrders
      tags: [lab]
      parameters:
        - { in: query, name: patientId, schema: { type: integer } }
        - { in: query, name: status, schema: { type: string } }
      responses:
        "200":
          description: Lab orders
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/LabOrder" }
    post:
      operationId: createLabOrder
      tags: [lab]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabOrderInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabOrder" }
  /lab/orders/{id}/result:
    post:
      operationId: recordLabResult
      tags: [lab]
      parameters:
        - { in: path, name: id, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabResultInput" }
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LabOrder" }
  /lab/orders/{id}/collect:
    post:
      operationId: collectLabSample
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectLabSampleInput" }
      responses:
        "200": { description: Collected, content: { application/json: { schema: { $ref: "#/components/schemas/LabOrder" } } } }
  /lab/orders/{id}/reject:
    post:
      operationId: rejectLabSample
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RejectLabSampleInput" }
      responses:
        "200": { description: Rejected, content: { application/json: { schema: { $ref: "#/components/schemas/LabOrder" } } } }
  /lab/orders/{id}/verify:
    post:
      operationId: verifyLabResult
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VerifyLabResultInput" }
      responses:
        "200": { description: Verified, content: { application/json: { schema: { $ref: "#/components/schemas/LabOrder" } } } }
  /lab/orders/{id}/dispatch:
    post:
      operationId: dispatchLabReport
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DispatchInput" }
      responses:
        "200": { description: Dispatched, content: { application/json: { schema: { $ref: "#/components/schemas/LabOrder" } } } }
  /lab/catalog:
    get:
      operationId: listLabCatalog
      tags: [lab]
      responses:
        "200": { description: Items, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/LabCatalogItem" } } } } }
    post:
      operationId: createLabCatalogItem
      tags: [lab]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabCatalogInput" }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/LabCatalogItem" } } } }
  /lab/catalog/{id}:
    patch:
      operationId: updateLabCatalogItem
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LabCatalogInput" }
      responses:
        "200": { description: Updated, content: { application/json: { schema: { $ref: "#/components/schemas/LabCatalogItem" } } } }
  /pdf/lab-report/{id}:
    get:
      operationId: getLabReportPdf
      tags: [lab]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: PDF, content: { application/pdf: { schema: { type: string, format: binary } } } }
...
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SendNotificationInput" }
      responses:
        "201":
          description: Sent
          content:
            application/json:
              schema: { $ref: "#/components/schemas/NotificationLog" }
...
components:
  schemas:
...
    LabResultParam:
      type: object
      required: [name, value]
      properties:
        name: { type: string }
        value: { type: string }
        unit: { type: ["string", "null"] }
        flag: { type: ["string", "null"], description: "N (normal), H (high), L (low), A (abnormal)" }
        refRange: { type: ["string", "null"] }
        comment: { type: ["string", "null"] }
    LabRefByAgeSex:
      type: object
      required: [sex, low, high]
      properties:
        minAgeYears: { type: ["integer", "null"] }
        maxAgeYears: { type: ["integer", "null"] }
        sex: { type: string, enum: [M, F, A] }
        low: { type: number }
        high: { type: number }
    LabCatalogParam:
      type: object
      required: [name]
      properties:
        name: { type: string }
        unit: { type: ["string", "null"] }
        refLow: { type: ["number", "null"] }
        refHigh: { type: ["number", "null"] }
        refText: { type: ["string", "null"] }
        refByAgeSex:
          type: array
          items: { $ref: "#/components/schemas/LabRefByAgeSex" }
    LabCatalogItem:
      type: object
      required: [id, code, name, sampleType, price, gstRate, turnaroundHours, active, createdAt]
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        category: { type: ["string", "null"] }
        sampleType: { type: string }
        container: { type: ["string", "null"] }
        price: { type: number }
        gstRate: { type: number }
        hsn: { type: ["string", "null"] }
        turnaroundHours: { type: integer }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/LabCatalogParam" }
        active: { type: boolean }
        createdAt: { type: string }
    LabCatalogInput:
      type: object
      required: [code, name, sampleType]
      properties:
        code: { type: string }
        name: { type: string }
        category: { type: string }
        sampleType: { type: string }
        container: { type: string }
        price: { type: number }
        gstRate: { type: number }
        hsn: { type: string }
        turnaroundHours: { type: integer }
        parameters:
          type: array
          items: { $ref: "#/components/schemas/LabCatalogParam" }
        active: { type: boolean }
    LabOrder:
      type: object
      required: [id, patientId, patientName, testName, status, priority, createdAt]
      properties:
        id: { type: integer }
        patientId: { type: integer }
        patientName: { type: string }
        patientAge: { type: ["integer", "null"] }
        patientSex: { type: ["string", "null"] }
        catalogId: { type: ["integer", "null"] }
        testName: { type: string }
        category: { type: ["string", "null"] }
        priority: { type: string }
        status: { type: string }
        result: { type: ["string", "null"] }
        normalRange: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        orderedBy: { type: ["string", "null"] }
        billId: { type: ["integer", "null"] }
        sampleId: { type: ["string", "null"] }
        barcode: { type: ["string", "null"] }
        collectedBy: { type: ["string", "null"] }
        collectedAt: { type: ["string", "null"] }
        rejectionReason: { type: ["string", "null"] }
        results:
          type: array
          items: { $ref: "#/components/schemas/LabResultParam" }
        attachmentUrl: { type: ["string", "null"] }
        verifiedBy: { type: ["string", "null"] }
        verifiedAt: { type: ["string", "null"] }
        reportPdfUrl: { type: ["string", "null"] }
        dispatchedAt: { type: ["string", "null"] }
        dispatchedVia: { type: ["string", "null"] }
        patientAcknowledgedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        completedAt: { type: ["string", "null"] }
    LabOrderInput:
      type: object
      required: [patientId]
      properties:
        patientId: { type: integer }
        catalogId: { type: integer }
        testName: { type: string }
        category: { type: string }
        priority: { type: string, enum: [routine, urgent, stat] }
        orderedBy: { type: string }
        autoBill: { type: boolean, description: "If true and catalogId set, auto-create a bill line for this order" }
    CollectLabSampleInput:
      type: object
      properties:
        collectedBy: { type: string }
        sampleId: { type: string }
        barcode: { type: string }
    RejectLabSampleInput:
      type: object
      required: [reason]
      properties:
        reason: { type: string }
    LabResultInput:
      type: object
      properties:
        result: { type: string, description: "Legacy free-text result; prefer 'results' parameter grid" }
        normalRange: { type: string }
        notes: { type: string }
        results:
          type: array
          items: { $ref: "#/components/schemas/LabResultParam" }
        attachmentUrl: { type: string }
    VerifyLabResultInput:
      type: object
      required: [verifiedBy]
      properties:
        verifiedBy: { type: string }
    DispatchInput:
      type: object
      properties:
        via: { type: string, enum: [sms, whatsapp, email, both, none] }
...
    SendNotificationInput:
      type: object
      required: [eventKey, channel, patientId, variables]
      properties:
        eventKey: { type: string }
        channel: { type: string }
        patientId: { type: integer }
        recipientPhone: { type: string }
        variables:
          type: object
          additionalProperties: { type: string }
    NotificationLog:
      type: object
      required: [id, eventKey, channel, status, renderedBody, recipientPhone, sentAt]
      properties:
        id: { type: integer }
        patientId: { type: ["integer", "null"] }
        patientName: { type: ["string", "null"] }
        eventKey: { type: string }
        channel: { type: string }
        templateId: { type: ["integer", "null"] }
        renderedBody: { type: string }
        recipientPhone: { type: string }
        status: { type: string }
        providerRef: { type: ["string", "null"] }
        errorMessage: { type: ["string", "null"] }
        sentAt: { type: string }
...